nits(components/nvidia): move packages in query
