+++ components/accelerator/nvidia/query/infiniband/infiniband_test.go
@@ -1,4 +1,4 @@
-package query
+package infiniband
 
 import "testing"
