+++ components/accelerator/nvidia/query/peermem/peermem_test.go
@@ -1,4 +1,4 @@
-package query
+package peermem
 
 import "testing"
